Issue fixes from v128(Kellogg's DSD)

+++ ivycpg/src/main/java/com/ivy/sd/png/provider/TaxHelper.java
@@ -831,8 +831,8 @@ public class TaxHelper implements TaxInterface {
     }
 
     @Override
-    public float updateProductWiseIncludeTax(List<ProductMasterBO> productMasterBOS) {
-        float totalTaxAmount = 0;
+    public double updateProductWiseIncludeTax(List<ProductMasterBO> productMasterBOS) {
+        double totalTaxAmount = 0;
         if (productMasterBOS != null && productMasterBOS.size() > 0) {
             for (ProductMasterBO productMasterBO : productMasterBOS) {
                 ProductMasterBO productBo = productMasterBO;
@@ -845,10 +845,10 @@ public class TaxHelper implements TaxInterface {
 
                             ArrayList<TaxBO> taxList = mTaxListByProductId.get(productBo.getProductID());
                             if (taxList != null) {
-                                float taxAmount = 0;
+                                double taxAmount = 0;
                                 for (TaxBO taxBO : taxList) {
                                     if (taxBO.getParentType().equals("0")) {
-                                        float calTax = SDUtil.truncateDecimal(productBo.getDiscount_order_value() * (taxBO.getTaxRate() / 100), 2).floatValue();
+                                        double calTax = SDUtil.truncateDecimal(productBo.getDiscount_order_value() * (taxBO.getTaxRate() / 100), 2).floatValue();
                                         taxBO.setTotalTaxAmount(calTax);
                                         taxAmount += calTax;
                                     }
